requirements.txt
azure_storage==0.37.0
Flask==3.0.3
Flask_Cors==4.0.1
flask_sqlalchemy==3.1.1
numpy==1.24.1
opencv_python==4.10.0.82
pandas==1.4.2
Pillow==9.1.0
Pillow==10.4.0
python-dotenv==1.0.1
scikit_learn==1.1.0
SQLAlchemy==2.0.30
tensorflow==2.8.0
~pencv_python==4.6.0.66
~pencv_python==4.7.0.68
